+++ angular.json
@@ -20,7 +20,7 @@
             "tsConfig": "tsconfig.app.json",
             "aot": true,
             "assets": [
-              "src/favicon.ico",
+              "src/assets/img/web-programming.svg",
               "src/assets",
               {
                 "glob": "**/*",
@@ -92,7 +92,7 @@
             "tsConfig": "tsconfig.spec.json",
             "karmaConfig": "karma.conf.js",
             "assets": [
-              "src/favicon.ico",
+              "src/assets/img/web-programming.svg",
               "src/assets"
             ],
             "styles": [
