+++ src/app/core/core.module.ts
@@ -10,6 +10,7 @@ import {AppRoutingModule} from '../app-routing.module';
 import {SidebarComponent} from './layout/sidebar/sidebar.component';
 import { PageNotFoundComponent } from './layout/page-not-found/page-not-found.component';
 import { PageUnderConstructionComponent } from './layout/page-under-construction/page-under-construction.component';
+import {IconModule} from '@ant-design/icons-angular';
 
 @NgModule({
   declarations: [FooterComponent, HeaderComponent, TranslateComponent, SidebarComponent, PageNotFoundComponent,
@@ -26,7 +27,8 @@ import { PageUnderConstructionComponent } from './layout/page-under-construction
     CommonModule,
     FormsModule,
     TranslateModule,
-    AppRoutingModule
+    AppRoutingModule,
+    IconModule
   ]
 })
 export class CoreModule {
