--- src/app/core/enum/MySocialNetworksEnum.ts
@@ -1,3 +1,0 @@
-export enum MySocialNetworksEnum {
-  LinkedIn = 'https://www.linkedin.com/in/alejandro-bolano/'
-}
