+++ src/app/core/enum/my-social-networks.enum.ts
@@ -1,0 +1,6 @@
+export enum MySocialNetworksEnum {
+  LinkedIn = 'https://www.linkedin.com/in/alejandro-bolano/',
+  Instagram = 'https://instagram.com/kubanox92',
+  Facebook = 'https://facebook.com/kubanox92',
+  Twitter = 'https://twitter.com/kubanox92'
+}
