+++ src/app/core/layout/sidebar/sidebar.component.html
@@ -1,7 +1,7 @@
 <div class="sidebar-logo">
   <a>
     <img src="{{imgLogo}}" alt="logo">
-    <h1>{{title}}</h1>
+    <h1>{{'titleSidebar'|translate}}</h1>
   </a>
 </div>
 <ul nz-menu nzTheme="{{theme}}" nzMode="inline" [nzInlineCollapsed]="isCollapsed">
@@ -21,7 +21,7 @@
       </li>
     </ul>
   </li>
-  <li nz-submenu nzOpen nzTitle="{{'private.title'|translate}}" nzIcon="lock">
+  <li *ngIf="isLoggedIn" nz-submenu nzOpen nzTitle="{{'private.title'|translate}}" nzIcon="lock">
     <ul>
       <li nz-menu-item nzMatchRouter (click)="collapseEvent()">
         <a routerLink="/private/sites">{{'private.sites'|translate}}</a>
@@ -37,5 +37,23 @@
         <a href="https://ng.ant.design/" target="_blank">Ant Design Of Angular</a>
       </li>
     </ul>
+  </li>
+  <li *ngIf="!isLoggedIn" nz-menu-item nzMatchRouter (click)="collapseEvent()">
+    <a routerLink="/authentication/log-in">
+      <i nz-icon nzType="login" nzTheme="outline"></i>
+      <span *ngIf="!isCollapsed">Log in</span>
+    </a>
+  </li>
+  <li *ngIf="isLoggedIn" nz-menu-item nzMatchRouter (click)="collapseEvent()">
+    <a routerLink="/authentication/dashboard">
+      <i nz-icon nzType="user" nzTheme="outline"></i>
+      <span *ngIf="!isCollapsed">{{userDataLoggedIn.email}}</span>
+    </a>
+  </li>
+  <li *ngIf="isLoggedIn" nz-menu-item nzMatchRouter (click)="collapseEvent()">
+    <a (click)="logOut()">
+      <i nz-icon nzType="logout" nzTheme="outline"></i>
+      <span *ngIf="!isCollapsed">Log out</span>
+    </a>
   </li>
 </ul>
