+++ src/app/pages/navigation/navigation.component.ts
@@ -2,8 +2,7 @@ import {Component, OnInit} from '@angular/core';
 
 @Component({
   selector: 'ambm-navigation',
-  template: `
-      <router-outlet></router-outlet>`
+  template: `<router-outlet></router-outlet>`
 })
 export class NavigationComponent implements OnInit {
