+++ src/assets/i18n/ca.json
@@ -1,6 +1,7 @@
 {
   "fullscreen": "Pantalla completa",
   "backHome": "Torna al principi",
+  "titleSidebar": "Alejandro M. Bolaño M.",
   "error": {
     "title": "Houston, tenim un problema!",
     "subtitle": "Ho sentim! la pàgina que intentes visitar no existeix."
