+++ src/assets/i18n/en.json
@@ -1,6 +1,7 @@
 {
   "fullscreen": "Fullscreen",
   "backHome": "Back home",
+  "titleSidebar": "Alejandro M. Bolaño M.",
   "error": {
     "title": "Houston, we have a problem!",
     "subtitle": "We are sorry! the page you visited does not exist."
