+++ src/assets/i18n/es.json
@@ -1,6 +1,7 @@
 {
   "fullscreen": "Pantalla completa",
   "backHome": "Regresa al principio",
+  "titleSidebar": "Alejandro M. Bolaño M.",
   "error": {
     "title": "¡Houston, tenemos un problema!",
     "subtitle": "¡Lo sentimos! la página que intentas visitar no existe."
