+++ src/index.html
@@ -5,7 +5,7 @@
   <title>Alejandro Bolaño</title>
   <base href="/">
   <meta name="viewport" content="width=device-width, initial-scale=1">
-  <link rel="icon" type="image/x-icon" href="favicon.ico">
+  <link rel="icon" type="image/x-icon" href="assets/img/web-programming.svg">
 </head>
 <body>
   <ambm-root></ambm-root>
